Run QUBO solvers using pathfinder graph partitioning and floating copy numbers

+++ tangle_resolution_benchmark_stats.sh
@@ -35,7 +35,8 @@ echo "===============" >> "$avg_file_name"
 
 
 
-for solver in $solvers; do    
+for solver in $solvers; do
+    [[ " pathfinder " =~ $solver ]] && continue;   
     avg_file_name="$solver.$annotate.$data_type.avg.txt"
     rm "$avg_file_name" 2> /dev/null
