+++ tangle_resolution_benchmark_violin.sh
@@ -6,6 +6,7 @@ data_type="$3"
 for annotate in km mg ga; do
 
     for solver in $solvers; do
+    [[ " pathfinder " =~ $solver ]] && continue;   
     rm "$solver.$annotate.$data_type.violin.txt" 2> /dev/null
     for t in ${time_limits//,/ }; do
         {
